requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
behave==1.2.6
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
outcome==1.3.0.post0
packaging==23.2
pandocfilters==1.5.1
parse==1.20.1
parse-type==0.6.2
parso==0.8.3
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
playwright==1.44.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.21
pyee==11.1.0
pygame==2.5.2
Pygments==2.17.2
PySocks==1.7.1
pytest==8.0.2
pytest-base-url==2.1.0
pytest-playwright==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.1.0
vobject==0.9.6.1
wcwidth==0.2.13
webdriver-manager==4.0.1
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
yarg==0.1.9
